Move pointcloud undistortion code into ROS-agnostic library
