+++ interfaces/ros1/wavemap_ros/CMakeLists.txt
@@ -24,7 +24,7 @@ cs_add_library(${PROJECT_NAME}
     src/map_operations/map_ros_operation_factory.cc
     src/map_operations/publish_map_operation.cc
     src/map_operations/publish_pointcloud_operation.cc
-    src/utils/pointcloud_undistortion/pointcloud_undistorter.cc
+    src/utils/pointcloud_undistorter.cc
     src/utils/rosbag_processor.cc
     src/utils/tf_transformer.cc
     src/ros_server.cc)
