+++ interfaces/ros1/wavemap_ros/include/wavemap_ros/utils/pointcloud_undistorter.h
@@ -1,5 +1,5 @@
-#ifndef WAVEMAP_ROS_UTILS_POINTCLOUD_UNDISTORTION_POINTCLOUD_UNDISTORTER_H_
-#define WAVEMAP_ROS_UTILS_POINTCLOUD_UNDISTORTION_POINTCLOUD_UNDISTORTER_H_
+#ifndef WAVEMAP_ROS_UTILS_POINTCLOUD_UNDISTORTER_H_
+#define WAVEMAP_ROS_UTILS_POINTCLOUD_UNDISTORTER_H_
 
 #include <algorithm>
 #include <memory>
@@ -9,8 +9,8 @@
 
 #include <wavemap/core/common.h>
 #include <wavemap/core/data_structure/pointcloud.h>
+#include <wavemap/core/utils/undistortion/stamped_pointcloud.h>
 
-#include "wavemap_ros/utils/pointcloud_undistortion/stamped_pointcloud.h"
 #include "wavemap_ros/utils/tf_transformer.h"
 
 namespace wavemap {
@@ -29,9 +29,10 @@ class PointcloudUndistorter {
         num_interpolation_intervals_per_cloud_(
             num_interpolation_intervals_per_cloud) {}
 
-  Result undistortPointcloud(StampedPointcloud& stamped_pointcloud,
-                             PosedPointcloud<>& undistorted_pointcloud,
-                             const std::string& fixed_frame);
+  Result undistortPointcloud(
+      undistortion::StampedPointcloud& stamped_pointcloud,
+      PosedPointcloud<>& undistorted_pointcloud,
+      const std::string& fixed_frame);
 
  private:
   std::shared_ptr<TfTransformer> transformer_;
@@ -39,4 +40,4 @@ class PointcloudUndistorter {
 };
 }  // namespace wavemap
 
-#endif  // WAVEMAP_ROS_UTILS_POINTCLOUD_UNDISTORTION_POINTCLOUD_UNDISTORTER_H_
+#endif  // WAVEMAP_ROS_UTILS_POINTCLOUD_UNDISTORTER_H_
