+++ library/include/wavemap/core/utils/undistortion/timestamps.h
@@ -1,0 +1,9 @@
+#ifndef WAVEMAP_CORE_UTILS_UNDISTORTION_TIMESTAMPS_H_
+#define WAVEMAP_CORE_UTILS_UNDISTORTION_TIMESTAMPS_H_
+
+namespace wavemap::undistortion {
+using TimeAbsolute = uint64_t;
+using TimeOffset = uint32_t;
+}  // namespace wavemap::undistortion
+
+#endif  // WAVEMAP_CORE_UTILS_UNDISTORTION_TIMESTAMPS_H_
