+++ library/src/core/CMakeLists.txt
@@ -37,6 +37,8 @@ target_sources(wavemap_core PRIVATE
     utils/sdf/full_euclidean_sdf_generator.cc
     utils/sdf/quasi_euclidean_sdf_generator.cc
     utils/time/stopwatch.cc
+    utils/undistortion/pointcloud_undistortion.cc
+    utils/undistortion/stamped_pointcloud.cc
     utils/thread_pool.cc)
 
 set_wavemap_target_properties(wavemap_core)
